+++ packages/app-renderer/src/MainWindow/index.tsx
@@ -1,9 +1,7 @@
-import { Provider as ReduxProvider } from 'react-redux';
 import { RouterProvider, createHashRouter } from 'react-router';
 
 import { FramelessWindow, TitleBar } from 'src/utils/frameless';
 
-import { mainStore } from './redux';
 import { routes } from './router/routes';
 
 const router = createHashRouter(routes);
@@ -12,9 +10,7 @@ export function MainWindow(): React.ReactNode {
     return (
         <FramelessWindow>
             <TitleBar>
-                <ReduxProvider store={mainStore}>
-                    <RouterProvider router={router} />
-                </ReduxProvider>
+                <RouterProvider router={router} />
             </TitleBar>
         </FramelessWindow>
     );