+++ packages/app-renderer/src/MainWindow/redux/hooks/use.app.ts
@@ -1,3 +1,0 @@
-import { useMainSelector } from './useSelector';
-
-export const useAppDetails = () => useMainSelector((state) => state.appDetails);