+++ packages/app-renderer/src/MainWindow/redux/hooks/useDispatching.ts
@@ -1,7 +1,0 @@
-import { useDispatchingActions, useDispatchingThunks } from 'src/utils/redux';
-
-import { actions } from '../actions';
-import { thunks } from '../thunks';
-
-export const useMainDispatchingActions = () => useDispatchingActions(actions);
-export const useMainDispatchingThunks = () => useDispatchingThunks(thunks);