+++ packages/app-renderer/src/MainWindow/redux/hooks/useSelector.ts
@@ -1,6 +1,0 @@
-import { useSelector } from 'react-redux';
-import type { TypedUseSelectorHook } from 'react-redux';
-
-import type { StoreState } from '../types';
-
-export const useMainSelector: TypedUseSelectorHook<StoreState> = useSelector;