+++ packages/app-renderer/src/MainWindow/redux/index.ts
@@ -1,5 +1,0 @@
-export { actions as mainActions, store as mainStore } from './actions';
-export { thunks as mainThunks } from './thunks';
-export type { StoreState as MainStoreState } from './types';
-
-export * from './hooks';