+++ packages/app-renderer/src/MainWindow/redux/thunks/createThunk.ts
@@ -1,5 +1,0 @@
-import { thunkCreatorFactory } from 'src/utils/redux';
-
-import type { StoreState } from '../types';
-
-export const createMainThunk = thunkCreatorFactory<StoreState>();