+++ packages/app-renderer/src/MainWindow/redux/thunks/index.ts
@@ -1,5 +1,0 @@
-import { prepareAppDetails } from './thunks.app';
-
-export const thunks = {
-    prepareAppDetails,
-};